// src/app/models/expenses.model.ts
export interface Expense {
  id: string;
  MonthName: string;
  Supermarket: number;
  Medical: number;
  Personal: number;
  EatingOut: number;
  Utilities: number;
  Takeaway: number;
  Tickets: number;
  Gaming: number;
  Cats: number;
  Travel: number;
  Subscriptions: number;
  Gym: number;
  TotalWage: number;
}